requirements.txt
Flask
docxtpl
python-docx
docx2pdf ; Windows uniquement
gunicorn